# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  university_system_internal_domain_models.Manager:
    properties:
      birthdate:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      department:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  university_system_internal_domain_models.Mark:
    properties:
      comment:
        type: string
      course_id:
        type: integer
      created_at:
        type: string
      final_mark:
        type: number
      first_attestation:
        type: number
      id:
        type: integer
      second_attestation:
        type: number
      student_id:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: n_rakhat@kbtu.kz
    name: Rakhat Nurtileu
    url: http://linkedin.com/in/nurtileu-rakhat-02aba3294/
  description: System для управления пользователями, курсами и ролями
  title: University System
  version: "1.0"
paths:
...
  /managers:
    get:
      description: Возвращает список всех менеджеров из базы данных.
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список менеджеров
          schema:
            items:
              $ref: '#/definitions/university_system_internal_domain_models.Manager'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить всех менеджеров
      tags:
      - managers
    post:
      consumes:
      - application/json
      description: Добавляет нового менеджера в систему.
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные нового менеджера
        in: body
        name: manager
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.Manager'
      produces:
      - application/json
      responses:
        "201":
          description: Созданный менеджер
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Manager'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Создать менеджера
      tags:
      - managers
    put:
      consumes:
      - application/json
      description: Изменяет данные менеджера.
      parameters:
      - description: Обновленные данные менеджера
        in: body
        name: manager
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.Manager'
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный менеджер
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Manager'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Обновить менеджера
      tags:
      - managers
  /managers/{id}:
    delete:
      description: Удаляет менеджера из системы по его идентификатору.
      parameters:
      - description: ID менеджера
        in: path
        name: id
        required: true
        type: string
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Удалить менеджера
      tags:
      - managers
    get:
      description: Возвращает данные менеджера по его идентификатору.
      parameters:
      - description: ID менеджера
        in: path
        name: id
        required: true
        type: string
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные менеджера
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Manager'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Менеджер не найден
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить менеджера по ID
      tags:
      - managers
  /managers/assign/{teacher_id}/{course_id}:
    post:
      description: Привязывает преподавателя к курсу.
      parameters:
      - description: ID преподавателя
        in: path
        name: teacher_id
        required: true
        type: string
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID курса
        in: path
        name: course_id
        required: true
        type: string
      responses:
        "200":
          description: Успешное назначение
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Назначить преподавателя на курс
      tags:
      - managers
  /marks/course/{course_id}:
    get:
      description: Возвращает все оценки студентов по заданному курсу.
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID курса
        in: path
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Mark'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить оценки по курсу
      tags:
      - marks
  /marks/student/{student_id}:
    get:
      description: Возвращает все оценки студента по всем курсам.
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID студента
        in: path
        name: student_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Mark'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить оценки студента
      tags:
      - marks
...
  /teachers/{id}/courses/{course_id}/students/{student_id}/PutFinalMark:
    post:
      consumes:
      - application/json
      description: Добавляет итоговую оценку студенту по курсу
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID учителя
        in: path
        name: id
        required: true
        type: string
      - description: ID студента
        in: path
        name: student_id
        required: true
        type: string
      - description: ID курса
        in: path
        name: course_id
        required: true
        type: string
      - description: Оценка
        in: body
        name: mark
        required: true
        schema:
          type: number
      produces:
      - application/json
      responses:
        "201":
          description: Оценка добавлена
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Ошибка ввода
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Запрещено
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Добавить итоговую оценку
      tags:
      - teachers
  /teachers/{id}/courses/{course_id}/students/{student_id}/PutFirstAtt:
    post:
      consumes:
      - application/json
      description: Добавляет первую аттестацию студенту по курсу
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID учителя
        in: path
        name: id
        required: true
        type: string
      - description: ID студента
        in: path
        name: student_id
        required: true
        type: string
      - description: ID курса
        in: path
        name: course_id
        required: true
        type: string
      - description: Оценка
        in: body
        name: mark
        required: true
        schema:
          type: number
      produces:
      - application/json
      responses:
        "201":
          description: Оценка добавлена
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Ошибка ввода
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Запрещено
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Добавить первую аттестацию
      tags:
      - teachers
  /teachers/{id}/courses/{course_id}/students/{student_id}/PutSecondAtt:
    post:
      consumes:
      - application/json
      description: Добавляет вторую аттестацию студенту по курсу
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID учителя
        in: path
        name: id
        required: true
        type: string
      - description: ID студента
        in: path
        name: student_id
        required: true
        type: string
      - description: ID курса
        in: path
        name: course_id
        required: true
        type: string
      - description: Оценка
        in: body
        name: mark
        required: true
        schema:
          type: number
      produces:
      - application/json
      responses:
        "201":
          description: Оценка добавлена
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Ошибка ввода
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Запрещено
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Добавить вторую аттестацию
      tags:
      - teachers
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"